requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bleach==6.2.0
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.0
Django==5.1.7
django-appconf==1.0.6
django-compressor==4.4
django-libsass==0.9
django-markdownify==0.9.5
django-tempus-dominus==5.1.2.17
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
frozenlist==1.4.0
gevent==23.9.1
geventhttpclient==2.0.2
greenlet==3.0.0
grpcio==1.59.0
gunicorn==23.0.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
libsass==0.23.0
Markdown==3.7
multidict==6.0.4
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
plotly==5.22.0
pluggy==1.5.0
propcache==0.2.0
protobuf==3.20.3
psycopg==3.1.12
psycopg-binary==3.1.12
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
PySocks==1.7.1
pytest==8.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-rapidjson==1.12
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.32.2
rjsmin==1.2.1
selenium==4.20.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
tenacity==8.3.0
tinycss2==1.4.0
trio==0.25.0
trio-websocket==0.11.1
tritonclient==2.36.0
typing_extensions==4.11.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.2.2
webdriver-manager==4.0.1
webencodings==0.5.1
whitenoise==6.5.0
wsproto==1.2.0
yarl==1.17.2
zope.event==5.0
zope.interface==6.1
